+++ lib/rhy-game.ts
@@ -43,26 +43,26 @@ abstract class Note {
     public judgement: 'none' | Judgement
     public count: number
 
-    private noteDOM: HTMLElement
+    public readonly DOM: HTMLElement
 
     public createDOM(laneDOM: HTMLBodyElement, moveTime: number, sizePerBeat: string, laneSizeRatio: number) {
         if (this.hasJudged) return
 
         for (const className of this.classNames) {
-            const currentClass = this.noteDOM.getAttribute('class') ?? ''
-            this.noteDOM.setAttribute('class', currentClass + ' ' + className)
+            const currentClass = this.DOM.getAttribute('class') ?? ''
+            this.DOM.setAttribute('class', currentClass + ' ' + className)
         }
-        this.noteDOM.style.setProperty('--size', `calc(${sizePerBeat} * ${this.sizeRatio})`)
-        this.noteDOM.style.animation = `${moveTime}ms ${this.timingFunction} ${this.moveAnimation}`
+        this.DOM.style.setProperty('--size', `calc(${sizePerBeat} * ${this.sizeRatio})`)
+        this.DOM.style.animation = `${moveTime}ms ${this.timingFunction} ${this.moveAnimation}`
 
-        this.noteDOM.addEventListener('animationend', () => {
-            this.noteDOM.style.animation = `${moveTime / laneSizeRatio * this.sizeRatio}ms ${this.timingFunction} ${this.fadeAnimation}`
-            this.noteDOM.addEventListener('animationend', () => {
-                this.noteDOM.remove()
+        this.DOM.addEventListener('animationend', () => {
+            this.DOM.style.animation = `${moveTime / laneSizeRatio * this.sizeRatio}ms ${this.timingFunction} ${this.fadeAnimation}`
+            this.DOM.addEventListener('animationend', () => {
+                this.DOM.remove()
             })
         })
 
-        laneDOM.appendChild(this.noteDOM)
+        laneDOM.appendChild(this.DOM)
     }
 
     public judge(judgements: Judgement[], eventName: EventName, actualTime: number) {
@@ -105,7 +105,7 @@ abstract class Note {
         this.judgement = 'none'
         this.count = 1
 
-        this.noteDOM = document.createElement('div')
+        this.DOM = document.createElement('div')
     }
 }